move ReflectiveCurveTo to separate file
