+++ core/src/commonMain/kotlin/dev/lennartegb/vec2compose/core/imagevector/Commands.kt
@@ -6,6 +6,7 @@ import dev.lennartegb.vec2compose.core.commands.Command
 import dev.lennartegb.vec2compose.core.commands.CurveTo
 import dev.lennartegb.vec2compose.core.commands.LineTo
 import dev.lennartegb.vec2compose.core.commands.MoveTo
+import dev.lennartegb.vec2compose.core.commands.ReflectiveCurveTo
 
 internal val Command.method: String get() = "$methodName($methodParams)"
 
@@ -18,7 +19,7 @@ private val Command.methodName: String
         is LineTo -> if (isAbsolute) "lineTo" else "lineToRelative"
         is MoveTo -> if (isAbsolute) "moveTo" else "moveToRelative"
         is Command.QuadraticBezierTo -> if (isAbsolute) "quadTo" else "quadToRelative"
-        is Command.ReflectiveCurveTo -> if (isAbsolute) "reflectiveCurveTo" else "reflectiveCurveToRelative"
+        is ReflectiveCurveTo -> if (isAbsolute) "reflectiveCurveTo" else "reflectiveCurveToRelative"
         is Command.ReflectiveQuadraticBezierTo -> if (isAbsolute) "reflectiveQuadTo" else "reflectiveQuadToRelative"
         is Command.VerticalLineTo -> if (isAbsolute) "verticalLineTo" else "verticalLineToRelative"
     }
@@ -35,7 +36,7 @@ private val Command.methodParams: String
         is LineTo -> floatParams(x, y)
         is MoveTo -> floatParams(x, y)
         is Command.QuadraticBezierTo -> floatParams(x1, y1, x2, y2)
-        is Command.ReflectiveCurveTo -> floatParams(x1, y1, x2, y2)
+        is ReflectiveCurveTo -> floatParams(x1, y1, x2, y2)
         is Command.ReflectiveQuadraticBezierTo -> floatParams(x, y)
         is Command.VerticalLineTo -> "${y}f"
     }
